# perf: cache more semver ranges