**What's the problem this PR addresses?**

There are still some locations in Yarn that uses uncached semver ranges slowing down installs.

Continues https://github.com/yarnpkg/berry/pull/1738

**How did you fix it?**

Cache `semverUtils.satisfiesWithPrereleases` and ensure all places in the codebase using semver uses `semverUtils.validRange` for ranges

**Result**

Installing the Storybook repo

```diff
  ➤ YN0000: ┌ Resolution step
- ➤ YN0000: └ Completed in 1s 46ms
+ ➤ YN0000: └ Completed in 0s 881ms
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.