@arcanis 2021-05-13
Is there an eslint plugin we could use to prevent some functions (like `semver.valid`) from being used? 🤔