@merceyz 2021-05-13
There is indeed, we should be able to use https://eslint.org/docs/rules/no-restricted-syntax